Split the control panel file into multiple files
